python: [API-Breaking Change] Migrated uboot.py to a uboot/ subpackage
Upcoming changes would only add to the amount of code in uboot.py, which
has become a bit of a junk drawer for functionality very tightly coupled
to U-Boot specific data structures and behavior.

As such, now is the time to refactor a little bit before only adding
more chaos.

The uboot.py submodule has been converted to a subpackage organized
as follows:

 * depthcharge.uboot.board:      Platform and "board" configuration data
 * depthcharge.uboot.cmd_table:  Console command handler table inspection
 * depthcharge.uboot.env:        Environment (variable) functionality
 * depthcharge.uboot.jump_table: Exported jump table definitions

This breaks the (unstable) v0.1.x API. Below is a summary of the v0.1.x
items that have been removed, along with their corresponding replacements in
the (also unstable) v0.2.x API.

depthcharge.uboot.board:
 v0.1.x: depthcharge.uboot.bdinfo_dict()

                     Replaced by:

 v0.2.x: depthcharge.uboot.board.bdinfo_dict()

depthcharge.uboot.cmd_table:
 v0.1.x: depthcharge.uboot.cmdtbl_entry_to_bytes()

                     Replaced by:

 v0.2.x: depthcharge.uboot.cmd_table.entry_to_bytes()

depthcharge.uboot.env:
 v0.1.x: depthcharge.uboot.raw_environment_regex()
         depthcharge.uboot.raw_env_var_regex()
         depthcharge.uboot.expand_environment()
         depthcharge.uboot.expand_variable()
         depthcharge.uboot.parse_raw_environment()
         depthcharge.uboot.load_raw_environment()
         depthcharge.uboot.save_raw_environment()
         depthcharge.uboot.create_raw_environment()
         depthcharge.uboot.parse_environment()
         depthcharge.uboot.load_environment()
         depthcharge.uboot.save_environment()

                     Replaced by:

 v0.2.x: depthcharge.uboot.env.raw_regex()
         depthcharge.uboot.env.raw_var_regex()
         depthcharge.uboot.env.expand()
         depthcharge.uboot.env.expand_variable()
         depthcharge.uboot.env.parse_raw()
         depthcharge.uboot.env.load_raw()
         depthcharge.uboot.env.save_raw()
         depthcharge.uboot.env.create_raw()
         depthcharge.uboot.env.load()
         depthcharge.uboot.env.save()

depthcharge.uboot.jump_table:
 v0.1.x: depthcharge.uboot.jump_table_exports()
         depthcharge.uboot.find_jump_table()

                     Replaced by:

 v0.2.x: depthcharge.uboot.jump_table.exports()
         depthcharge.uboot.jump_table.find()
