+++ python/depthcharge/uboot/__init__.py
@@ -1,0 +1,13 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Depthcharge: <https://github.com/nccgroup/depthcharge>
+#
+# flake8: noqa=F401
+
+"""
+U-Boot centric parsing, conversion, and data processing functionality
+"""
+
+from . import board
+from . import cmd_table
+from . import env
+from . import jump_table
